requirements.txt
discord.py[voice]>=2.3.2
asyncpg>=0.29.0
alembic>=1.13.1
pydantic>=2.6.4
python-dotenv>=1.0.1
structlog>=24.1.0
